add ALB_PREDEFINED_SSL_POLICY_CHECK
