+++ python/IAM_USER_USED_LAST_90_DAYS/parameters.json
@@ -1,16 +1,12 @@
 {
   "Version": "1.0",
   "Parameters": {
-    "RuleName": "IAM_PRINCIPAL_USED_LAST_90_DAYS",
+    "RuleName": "IAM_USER_USED_LAST_90_DAYS",
     "SourceRuntime": "python3.6",
-    "CodeKey": "iamlastused.zip",
+    "CodeKey": "IAM_USER_USED_LAST_90_DAYS.zip",
     "InputParameters": "{}",
     "OptionalParameters": "{\"WhitelistedUserList\": \"\", \"NotUsedTimeOutInDays\": \"\", \"NewUserCooldownInDays\": \"\"}",
     "SourceEvents": "AWS::IAM::User",
-    "SourcePeriodic": "TwentyFour_Hours",
-    "RuleSets": [
-      "baseline",
-      "rulecriticity:medium"
-    ]
+    "SourcePeriodic": "TwentyFour_Hours"
   }
 }
